post article route
